cartoonize/requirements.txt
Flask==1.0.2
gunicorn==20.0.4
Pillow==6.2.0
opencv_python
tensorflow

google-cloud-storage==1.29.0
algorithmia==1.3.0
scikit-video==1.1.11
tf_slim==1.1.0
PyYaml==5.3.1
flask-ngrok